pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.1.2",
    "psycopg2-binary>=2.9.11",
    "pycryptodome>=3.23.0",
    "python-socketio>=5.14.2",
]
